chore: resolve feedback
